# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * KENKAI AI Strategic Thinking Partner API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type AssessmentType = typeof AssessmentType[keyof typeof AssessmentType];


export const AssessmentType = {
  personal: 'personal',
  career: 'career',
  business: 'business',
} as const;

export type AssessmentStatus = typeof AssessmentStatus[keyof typeof AssessmentStatus];


export const AssessmentStatus = {
  not_started: 'not_started',
  in_progress: 'in_progress',
  completed: 'completed',
} as const;

export interface Assessment {
  id: number;
  type: AssessmentType;
  title: string;
  /** @nullable */
  description?: string | null;
  status: AssessmentStatus;
  progress?: number;
  /** @nullable */
  completedAt?: string | null;
  createdAt: string;
}

export type AssessmentInputType = typeof AssessmentInputType[keyof typeof AssessmentInputType];


export const AssessmentInputType = {
  personal: 'personal',
  career: 'career',
  business: 'business',
} as const;

export interface AssessmentInput {
  type: AssessmentInputType;
  title: string;
  description?: string;
}

export type AssessmentUpdateStatus = typeof AssessmentUpdateStatus[keyof typeof AssessmentUpdateStatus];


export const AssessmentUpdateStatus = {
  not_started: 'not_started',
  in_progress: 'in_progress',
  completed: 'completed',
} as const;

export interface AssessmentUpdate {
  status?: AssessmentUpdateStatus;
  progress?: number;
}

export type SessionCategory = typeof SessionCategory[keyof typeof SessionCategory];


export const SessionCategory = {
  personal: 'personal',
  career: 'career',
  business: 'business',
} as const;

export type SessionStatus = typeof SessionStatus[keyof typeof SessionStatus];


export const SessionStatus = {
  active: 'active',
  completed: 'completed',
  archived: 'archived',
} as const;

export interface Session {
  id: number;
  title: string;
  category: SessionCategory;
  status: SessionStatus;
  messageCount: number;
  /** @nullable */
  lastMessage?: string | null;
  /** @nullable */
  assessmentId?: number | null;
  createdAt: string;
  updatedAt?: string;
}

export type SessionInputCategory = typeof SessionInputCategory[keyof typeof SessionInputCategory];


export const SessionInputCategory = {
  personal: 'personal',
  career: 'career',
  business: 'business',
} as const;

export interface SessionInput {
  title: string;
  category: SessionInputCategory;
  assessmentId?: number;
}

export type SessionWithMessagesCategory = typeof SessionWithMessagesCategory[keyof typeof SessionWithMessagesCategory];


export const SessionWithMessagesCategory = {
  personal: 'personal',
  career: 'career',
  business: 'business',
} as const;

export type SessionWithMessagesStatus = typeof SessionWithMessagesStatus[keyof typeof SessionWithMessagesStatus];


export const SessionWithMessagesStatus = {
  active: 'active',
  completed: 'completed',
  archived: 'archived',
} as const;

export type MessageRole = typeof MessageRole[keyof typeof MessageRole];


export const MessageRole = {
  user: 'user',
  assistant: 'assistant',
} as const;

export interface Message {
  id: number;
  sessionId: number;
  role: MessageRole;
  content: string;
  createdAt: string;
}

export interface SessionWithMessages {
  id: number;
  title: string;
  category: SessionWithMessagesCategory;
  status: SessionWithMessagesStatus;
  messageCount: number;
  /** @nullable */
  lastMessage?: string | null;
  /** @nullable */
  assessmentId?: number | null;
  createdAt: string;
  updatedAt: string;
  messages: Message[];
}

export interface MessageInput {
  content: string;
}

export type ReportType = typeof ReportType[keyof typeof ReportType];


export const ReportType = {
  personality: 'personality',
  leadership: 'leadership',
  business_health: 'business_health',
  customer_persona: 'customer_persona',
  strategic_recommendations: 'strategic_recommendations',
  life_plan: 'life_plan',
  career_roadmap: 'career_roadmap',
  action_plan: 'action_plan',
} as const;

export type ReportStatus = typeof ReportStatus[keyof typeof ReportStatus];


export const ReportStatus = {
  generating: 'generating',
  ready: 'ready',
} as const;

export interface Report {
  id: number;
  type: ReportType;
  title: string;
  summary: string;
  /** @nullable */
  content?: string | null;
  keyInsights?: string[];
  status: ReportStatus;
  /** @nullable */
  assessmentId?: number | null;
  /** @nullable */
  sessionId?: number | null;
  createdAt: string;
}

export type ReportGenerateInputType = typeof ReportGenerateInputType[keyof typeof ReportGenerateInputType];


export const ReportGenerateInputType = {
  personality: 'personality',
  leadership: 'leadership',
  business_health: 'business_health',
  customer_persona: 'customer_persona',
  strategic_recommendations: 'strategic_recommendations',
  life_plan: 'life_plan',
  career_roadmap: 'career_roadmap',
  action_plan: 'action_plan',
} as const;

export interface ReportGenerateInput {
  type: ReportGenerateInputType;
  title: string;
  assessmentId?: number;
  sessionId?: number;
}

export interface DashboardOverview {
  totalAssessments: number;
  completedAssessments: number;
  activeSessions: number;
  totalReports: number;
  currentStreak: number;
  clarityScore: number;
}

export type ActivityItemType = typeof ActivityItemType[keyof typeof ActivityItemType];


export const ActivityItemType = {
  assessment_completed: 'assessment_completed',
  session_started: 'session_started',
  report_generated: 'report_generated',
  milestone_reached: 'milestone_reached',
} as const;

export interface ActivityItem {
  id: number;
  type: ActivityItemType;
  title: string;
  description: string;
  createdAt: string;
}

export type ActionPlanTimeframe = typeof ActionPlanTimeframe[keyof typeof ActionPlanTimeframe];


export const ActionPlanTimeframe = {
  '30_days': '30_days',
  '60_days': '60_days',
  '90_days': '90_days',
} as const;

export type ActionPlanStatus = typeof ActionPlanStatus[keyof typeof ActionPlanStatus];


export const ActionPlanStatus = {
  active: 'active',
  completed: 'completed',
} as const;

export interface ActionPlan {
  id: number;
  title: string;
  timeframe: ActionPlanTimeframe;
  tasks: number;
  completedTasks: number;
  status: ActionPlanStatus;
  /** @nullable */
  reportId?: number | null;
}