# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: KENKAI AI Strategic Thinking Partner API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /assessments:
    get:
      operationId: listAssessments
      tags: [assessments]
      summary: List all assessments
      responses:
        "200":
          description: List of assessments
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Assessment"
    post:
      operationId: createAssessment
      tags: [assessments]
      summary: Create a new assessment
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AssessmentInput"
      responses:
        "201":
          description: Created assessment
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Assessment"
  /assessments/{id}:
    get:
      operationId: getAssessment
      tags: [assessments]
      summary: Get assessment by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Assessment
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Assessment"
        "404":
          description: Not found
    patch:
      operationId: updateAssessment
      tags: [assessments]
      summary: Update assessment
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AssessmentUpdate"
      responses:
        "200":
          description: Updated assessment
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Assessment"
...
components:
  schemas:
...
    Assessment:
      type: object
      required: [id, type, title, status, createdAt]
      properties:
        id:
          type: integer
        type:
          type: string
          enum: [personal, career, business]
        title:
          type: string
        description:
          type: ["string", "null"]
        status:
          type: string
          enum: [not_started, in_progress, completed]
        progress:
          type: integer
        completedAt:
          type: ["string", "null"]
        createdAt:
          type: string
    AssessmentInput:
      type: object
      required: [type, title]
      properties:
        type:
          type: string
          enum: [personal, career, business]
        title:
          type: string
        description:
          type: string
    AssessmentUpdate:
      type: object
      properties:
        status:
          type: string
          enum: [not_started, in_progress, completed]
        progress:
          type: integer
...