# lib/api-zod/src/generated/types/assessmentInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * KENKAI AI Strategic Thinking Partner API
 * OpenAPI spec version: 0.1.0
 */
import type { AssessmentInputType } from './assessmentInputType';

export interface AssessmentInput {
  type: AssessmentInputType;
  title: string;
  description?: string;
}