# lib/api-zod/src/generated/types/assessmentUpdate.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * KENKAI AI Strategic Thinking Partner API
 * OpenAPI spec version: 0.1.0
 */
import type { AssessmentUpdateStatus } from './assessmentUpdateStatus';

export interface AssessmentUpdate {
  status?: AssessmentUpdateStatus;
  progress?: number;
}